requirements.txt
requests>=2.31.0
requests[socks]>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6
ipaddress>=1.0.23
dnspython>=2.4.0
